fix: Address code review feedback for browserless integration
Critical fixes:
- Use json_encode() instead of addslashes() for JavaScript credential escaping
  to properly handle special characters, template literals, and Unicode
- Fix stale in-memory value bug in markSyncFailed() by computing new failure
  count before update and handling sync_enabled in single atomic operation
- Remove duplicate failure counting by only calling markSyncFailed() in
  failed() method after all retries are exhausted

Security improvements:
- Remove username from log statements to prevent PII exposure
- Add security warning comment in .env.example for portal credentials

Route improvements:
- Change test-connection route from GET to POST for RESTful compliance
- Update documentation to include all routes (toggle-auto-sync, test-connection)

Code quality improvements:
- Add integer casting to config values (timeout, retry attempts/delay)
- Remove unused $request parameter from BrowserlessController::history()

Files modified:
- app/Services/BrowserlessService.php - JavaScript escaping and logging
- app/Models/BrowserlessConnection.php - Atomic failure tracking
- app/Jobs/FetchInvestorData.php - Single failure counting point
- app/Http/Controllers/BrowserlessController.php - Remove unused param
- routes/web.php - Change test-connection to POST
- config/browserless.php - Add integer casting
- .env.example - Add security warning
- INVESTOR_PORTAL_INTEGRATION.md - Document all routes

+++ config/browserless.php
@@ -31,7 +31,7 @@ return [
     | Maximum time in seconds to wait for a crawl operation to complete.
     |
     */
-    'timeout' => env('BROWSERLESS_TIMEOUT', 60),
+    'timeout' => (int) env('BROWSERLESS_TIMEOUT', 60),
 
     /*
     |--------------------------------------------------------------------------
@@ -57,7 +57,7 @@ return [
     |
     */
     'retry' => [
-        'max_attempts' => env('BROWSERLESS_RETRY_ATTEMPTS', 3),
-        'delay_seconds' => env('BROWSERLESS_RETRY_DELAY', 60),
+        'max_attempts' => (int) env('BROWSERLESS_RETRY_ATTEMPTS', 3),
+        'delay_seconds' => (int) env('BROWSERLESS_RETRY_DELAY', 60),
     ],
 ];
